feat: created api-docs using swagger refactor!: moved files into new directories feat!: Created new route /api-docs for swagger automated documentation
