+++ package.json
@@ -22,7 +22,9 @@
     "esm": "^3.2.25",
     "express": "^4.17.1",
     "express-validator": "^6.10.0",
-    "mongoose": "^5.12.0"
+    "mongoose": "^5.12.0",
+    "swagger-jsdoc": "^6.1.0",
+    "swagger-ui-express": "^4.1.6"
   },
   "devDependencies": {
     "nodemon": "^2.0.7"
