--- routes/formData/formData.js
@@ -1,9 +1,0 @@
-import express from 'express';
-
-import { characterFormData } from '../../controllers/characterData/characterDataController';
-
-const router = express.Router();
-
-router.get('/character', characterFormData);
-
-export default router;
