+++ src/models/character/characterModel.js
